Interface to LibRI, and others

+++ src/chi0.h
@@ -17,13 +17,12 @@ using std::vector;
 class Chi0
 {
     private:
-        unsigned gf_save;
-        unsigned gf_discard;
+        size_t gf_save;
+        size_t gf_discard;
         //! space-time Green's function in occupied space, [ispin][I][J][R][tau]
         /*!
-         * @note: itau (index) less than zero correspond to occupied GF,
+         * @note: tau (index) less than zero correspond to occupied GF,
          *        and larger than zero correspond to unoccpued GF.
-         *        Absolute value starting from 1 to the size of tfg.
          * @note: May need to use ComplexMatrix for GF.
          */
         map<int, atom_mapping<map<Vector3_Order<int>, map<double, matrix>>>::pair_t_old> gf_is_R_tau;
